requirements.txt
fastapi==0.115.6
huggingface_hub==0.27.1
pandas==2.2.3
python-dotenv==1.0.1
